code restructuring
