+++ src/analyzing/index.ts
