+++ src/analyzing/type_inference/index.ts
@@ -1,0 +1,1 @@
+export { InferAccessType } from './InferAccessType'
