+++ src/analyzing/types/CurriedTypeConstructor.ts
