+++ src/analyzing/types/ListType.ts
