+++ src/analyzing/types/MapType.ts
