+++ src/analyzing/types/ObjectType.ts
