+++ src/analyzing/types/SingleTypeConstructor.ts
