+++ src/analyzing/types/Type.ts
