+++ src/analyzing/types/TypeConstructor.ts
