+++ src/analyzing/types/TypeInterface.ts
