+++ src/error_handling/index.ts
@@ -1,0 +1,1 @@
+export { ErrorHandler } from './ErrorHandler'
